создал компоненты Header.tsx и Technologies.tsx
