+++ src/Header.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+
+const Header = () => {
+    return (
+        <div>
+            <a href="#">Home</a>
+            <a href="#">New Feed</a>
+            <a href="#">Messages</a>
+        </div>
+    )
+
+}
+export default Header;
