+++ src/Technologies.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+
+
+const Technologies = () => {
+    return (
+        <div>
+            <h2>Hello, samurai! Let's go!</h2>
+            <ul>
+                <li>HTML</li>
+                <li>CSS</li>
+                <li>JS</li>
+                <li>React</li>
+            </ul>
+        </div>
+    )
+
+}
+
+export default Technologies;
